+++ lib/ui/page/home/page/settings/controller.dart
@@ -14,6 +14,7 @@
 // along with this program. If not, see
 // <https://www.gnu.org/licenses/agpl-3.0.html>.
 
+import 'package:flutter/widgets.dart' show GlobalKey;
 import 'package:get/get.dart';
 
 import '/domain/model/application_settings.dart';
@@ -23,6 +24,9 @@ import '/l10n/l10n.dart';
 /// Controller of the [Routes.settings] page.
 class SettingsController extends GetxController {
   SettingsController(this._settingsRepo);
+
+  /// [GlobalKey] of a button opening the [Language] selection.
+  final GlobalKey languageKey = GlobalKey();
 
   /// Settings repository, used to update the [ApplicationSettings].
   final AbstractSettingsRepository _settingsRepo;